DRY up the form error struct
closes 61
