+++ src/authenticated.rs
@@ -43,6 +43,40 @@ pub struct UserExtension {
     csrf: String,
 }
 
+#[derive(Debug)]
+pub struct FormError {
+    message: String,
+}
+
+impl IntoResponse for FormError {
+    fn into_response(self) -> Response {
+        return (StatusCode::BAD_REQUEST, format!("{:#?}", self)).into_response();
+    }
+}
+
+impl From<bson::oid::Error> for FormError {
+    fn from(value: bson::oid::Error) -> Self {
+        FormError {
+            message: value.to_string(),
+        }
+    }
+}
+
+impl From<tera::Error> for FormError {
+    fn from(value: tera::Error) -> Self {
+        FormError {
+            message: value.to_string(),
+        }
+    }
+}
+impl From<mongodb::error::Error> for FormError {
+    fn from(value: mongodb::error::Error) -> Self {
+        FormError {
+            message: value.to_string(),
+        }
+    }
+}
+
 async fn validate_csrf(
     user: Extension<UserExtension>,
     headers: HeaderMap,
